duinisveldbreugem/urls.py
from rest_framework import routers
from django.urls import path, include
from . import views


router = routers.DefaultRouter()
router.register(r'invoices', views.InvoiceViewSet)
router.register(r'invoice-items', views.InvoiceItemViewSet)


urlpatterns = [
    path('', include(router.urls)),
]