requirements.txt
asgiref==3.4.1
astroid==2.5.6
backports.entry-points-selectable==1.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
cfgv==3.3.1
charset-normalizer==2.0.7
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.8
django-filter==21.1
django-ratelimit==3.0.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dnspython==2.2.1
et-xmlfile==1.1.0
exchangelib==4.8.0
filelock==3.0.12
future==0.18.2
gunicorn==20.1.0
identify==2.2.14
idna==3.3
IMAPClient==2.3.1
isodate==0.6.1
isort==5.8.0
itypes==1.2.0
Jinja2==3.0.2
lazy-object-proxy==1.6.0
lxml==4.9.1
mail-parser==3.15.0
MarkupSafe==2.0.1
mccabe==0.6.1
nodeenv==1.6.0
ntlm-auth==1.5.0
numpy==1.21.3
O365==2.0.21
oauthlib==3.1.1
openpyxl==3.0.9
pandas==1.3.4
platformdirs==2.3.0
pre-commit==2.15.0
pycef==1.11
pycparser==2.20
Pygments==2.13.0
PyJWT==2.3.0
pylint==2.8.2
python-dateutil==2.8.2
python-dotenv==0.19.1
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
ratelimit==2.2.1
requests==2.26.0
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stringcase==1.2.0
toml==0.10.2
tzdata==2022.5
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.7
virtualenv==20.7.2
wrapt==1.12.1
xlrd==2.0.1
